server/requirements.txt
Flask
bcrypt
Flask-Cors
pyjwt
pymongo
mongoengine
marshmallow
pytest
mongomock